Adds support for GitLab CI

+++ source/runner/Executor.ts
@@ -251,7 +251,7 @@ export class Executor {
       await this.platform.deleteMainComment(dangerID)
       const previousComments = await this.platform.getInlineComments(dangerID)
       for (const comment of previousComments) {
-        if (comment) {
+        if (comment && comment.ownedByDanger) {
           await this.deleteInlineComment(comment)
         }
       }
@@ -277,6 +277,7 @@ export class Executor {
         this.platform.deleteMainComment(dangerID)
       } else {
         const commitID = git.commits[git.commits.length - 1].sha
+        // TODO: GitLab template formatting (or reuse one of the others?)
         const comment = process.env["DANGER_BITBUCKETSERVER_HOST"]
           ? bitbucketServerTemplate(dangerID, commitID, mergedResults)
           : githubResultsTemplate(dangerID, commitID, mergedResults)
